pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trip-planner"
version = "0.1.0"
description = ""
authors=[
  { name = "Tamir Bahar" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs",
    "rich",
    "diskcache",
    "simplekml",
    "httpx",
    "urllib3",
    "python-docx",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "isort",
    "nox",
    "mypy",
    "pytest",
    "pytest-cov",
]




[project.scripts]
doc2map = "trip_planner.trip_planner:app"




[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
# Configure isort to work without access to site-packages
known_first_party = ["trip_planner", "tests"]

# Settings for Black compatibility
profile = "black"
line_length = 88


[tool.mypy]
check_untyped_defs = true
python_version = "3.11"

[tool.mypy-diskcache]
ignore_missing_imports = true


[tool.pytest.ini_options]
xfail_strict = true
addopts = "-ra -q --cov --cov-report=xml --cov-report=term-missing"
testpaths = [
    "tests",
]


[tool.ruff.lint]
# Avoid automatic removal of unused variables
unfixable = ['F841'] # unused-variable


[tool.coverage]
report.skip_covered = true
